net: msm_rmnet.h: Fix data type of flow_id in QoS header
Fix an issue where the flow_id was appearing to be 8 bytes in
64 bit platforms. This is because the flow_id is declared as unsigned
long which is 8 bytes in 64 bit and 4 bytes in 32 bit platforms.
Change the flow_id to uint32_t to make it 4 bytes on all platforms.

CRs-fixed: 663814

+++ include/linux/msm_rmnet.h
@@ -37,7 +37,7 @@ enum rmnet_ioctl_cmds_e {
 struct QMI_QOS_HDR_S {
 	unsigned char    version;
 	unsigned char    flags;
-	unsigned long    flow_id;
+	uint32_t         flow_id;
 };
 
 #endif /* _MSM_RMNET_H_ */
